+++ lib/packages/core/CHANGELOG.md
@@ -2,5 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
-### Added
+### Network
+- Added `suggestFees` for getting suggested fees.
+- Added `sendMoney` for generating the unsigned transaction, signing it, and broadcasting it.
+
+### Account
 - Added `generateSendTransactionQRCode` and `generateSendTransactionQRCodeAddress` methods for generating a QR code image or URL, respectively.